finished ad

+++ src/main/java/com/codeup/adlister/dao/MySQLAdsDao.java
@@ -61,8 +61,8 @@ public class MySQLAdsDao implements Ads {
 
     private Ad extractAd(ResultSet rs) throws SQLException {
         return new Ad(
-            rs.getInt("id"),
-            rs.getInt("user_id"),
+            rs.getLong("id"),
+            rs.getLong("user_id"),
             rs.getString("title"),
             rs.getString("description"),
             rs.getFloat("cost"),
